fix: 구조 개선
