+++ apps/client/src/app/[user]/components/RecommendSection/RecommendList/index.css.ts
@@ -1,0 +1,6 @@
+import { style } from "@vanilla-extract/css";
+
+export const studyListWrapper = style({
+  display: "flex",
+  flexDirection: "column",
+});
